Re-structure the bank balance system: revert to original method of using the local *_usage files for finding usage instead of the SlurmDBD.
This reverts the change made in commit 4f42daf9

The reasoning for reverting back to the previous method is that
the balance of available time in slurm is found by subtracting
the usage value from the local *_usage files; so to get an accurate
balance we should use them. The SlurmDBD sreport reports are
still useful for reporting, but they do not reflect the notion
of available balance (and so the displayed balance is not accurate).
Also, the SlurmDBD sreport values can easily grow larger than the
available Limit, and hence leads to a display of a negative balance.

Note also that if a PriorityDecayHalfLife or PriorityUsageResetPeriod
are in use, they will now be reflected in the correct balance (since
they affect the local *_usage files and not the SlurmDBD).

Old system:
- read usage information from SlurmDBD via 'sreport'

New system:
- read usage information from local *_usage files via 'sshare'

TODO: update docs to reflect changes; make the previous SlurmDBD
usage report available via the '-s' flag.